requirements.txt
numpy>=1.14.0
gdown>=3.10.1
Pillow>=5.2.0
opencv-python>=3.4.4
tensorflow>=1.9.0
Click==7.0
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-dotenv==0.10.1
Werkzeug==0.14.1
WTForms==2.2.1